Extract metadata-based install to class
